backend/pyproject.toml
[project]
name = "zerg"
version = "0.1.0"
description = "AI Agent Platform Backend"
authors = [
    {name = "David Rose", email = "david.github@drose.io"}
]
requires-python = ">=3.12"
readme = "../README.md"
license = {text = "MIT"}

dependencies = [
    "python-dotenv>=1.0.0",
    "black",
    "ruff",
    "pytest",
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "websockets>=12.0",
    "openai>=1.65.2",
    "sqlalchemy>=2.0.38",
    "requests>=2.32.3",
    "pytest-xdist>=3.6.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["zerg"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["I", "E", "F", "Q000"]
ignore = []

[tool.ruff.lint.isort]
force-single-line = true

[tool.ruff.lint.flake8-quotes]
docstring-quotes = "double"
inline-quotes = "double"
multiline-quotes = "double"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

